+++ mitzasql/ui/widgets/table_info_widget.py
@@ -20,11 +20,9 @@
 
 import urwid
 
-import mitzasql.ui.utils as utils
+from mitzasql.ui.widgets.info_widget import InfoWidget
 
-class TableInfoWidget(urwid.ListBox):
-    SIGNAL_ESCAPE = 'escape'
-
+class TableInfoWidget(InfoWidget):
     def __init__(self, model):
         self._model = model;
         head_section = self._create_head_section()
@@ -33,19 +31,10 @@ class TableInfoWidget(urwid.ListBox):
         contents = [head_section]
         contents.extend(body_section)
         super().__init__(contents)
-        urwid.register_signal(self.__class__, [self.SIGNAL_ESCAPE])
 
     @property
     def name(self):
         return u'Table {0}'.format(self._model['table']['TABLE_NAME'])
-
-    def keypress(self, size, key):
-        key = utils.vim2emacs_translation(key)
-        if key == 'esc':
-            urwid.emit_signal(self, self.SIGNAL_ESCAPE, self)
-            return
-
-        return super().keypress(size, key)
 
     def _create_head_section(self):
         grid = []