+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update user manual
 
 ### Fixed
+- Fix for #65: Catch exit key in info widgets
 - Fix `end` key scrolling issue (#75)
 - Fix connect using command line arguments bug